Update integer width to avoid overflow

+++ firmware/Core/Src/rtos_tasks/rtos_tasks.c
@@ -125,7 +125,7 @@ void TASK_handle_uart_telecommands(void *argument) {
 
 			// TODO: maybe log/print args too
 			
-			const uint8_t arg_len = end_of_args_idx - start_of_args_idx - 1;
+			const uint16_t arg_len = end_of_args_idx - start_of_args_idx - 1;
 			uint8_t args_str_no_parens[arg_len + 1];
 			memcpy(args_str_no_parens, &latest_tcmd[start_of_args_idx + 1], arg_len);
 			args_str_no_parens[arg_len] = '\0';
